Add DSL extension feature (default shipped with sudo execution and file transfer)
