+++ src/main/groovy/org/hidetake/gradle/ssh/internal/session/SessionService.groovy
@@ -1,5 +1,6 @@
 package org.hidetake.gradle.ssh.internal.session
 
+import groovy.util.logging.Slf4j
 import org.hidetake.gradle.ssh.internal.connection.ConnectionManager
 import org.hidetake.gradle.ssh.internal.operation.OperationService
 import org.hidetake.gradle.ssh.plugin.OperationSettings
@@ -7,10 +8,15 @@ import org.hidetake.gradle.ssh.plugin.Remote
 import org.hidetake.gradle.ssh.plugin.session.SessionHandler
 
 @Singleton(lazy = true)
+@Slf4j
 class SessionService {
     SessionHandler createDelegate(Remote remote, OperationSettings operationSettings, ConnectionManager connectionManager) {
         def operationService = OperationService.instance
         def operations = operationService.create(remote, operationSettings, connectionManager)
-        new DefaultSessionHandler(operations, operationSettings)
+
+        def handler = new DefaultSessionHandler(operations, operationSettings)
+        log.info("Mixin extensions: ${operationSettings.extensions}")
+        handler.metaClass.mixin(operationSettings.extensions)
+        handler
     }
 }
